# Avoided the need for prefiltering the ruptures twice