--- a/openquake/calculators/base.py
+++ b/openquake/calculators/base.py
@@ -733,11 +733,6 @@ class HazardCalculator(BaseCalculator):
 
         # compute exposure stats
         if hasattr(self, 'assetcol'):
-            arr = self.assetcol.array
-            num_assets = list(general.countby(arr, 'site_id').values())
-            self.datastore['assets_by_site'] = get_stats(num_assets)
-            num_taxos = self.assetcol.num_taxonomies_by_site()
-            self.datastore['taxonomies_by_site'] = get_stats(num_taxos)
             save_exposed_values(
                 self.datastore, self.assetcol, oq.loss_names, oq.aggregate_by)