Thanks to the change in https://github.com/gem/oq-engine/pull/5369 it is now finally possible to speedup `RuptureGetter.get_ruptures` significantly by making use of the prefiltering already performed in the master node. For instance in the case of Chile there is a 5x speedup resulting in nearly halving the total runtime.
```
# master
total ebrisk                                70,079   2,740     212
getting ruptures                            33,918   0.0       1,159,570
getting hazard                              19,272   0.0       1,144,859
aggregating losses                          14,747   0.0       114,454

# this branch
total ebrisk                                40,926   2,801     212      
getting hazard                              18,588   0.0       1,144,859
aggregating losses                          13,811   0.0       114,454  
getting ruptures                            6,564    0.0       1,159,570
```
The cost is a much larger data transfer in the RuptureGetter (from 83 MB to 1.47 GB, nearly 20x) but it is a non-issue now that we have zmq.